Reuse stance indicator UI in debate now modal
